requirements.txt
openai
flask
boto3
requests
Pillow